+++ example-app/models/Book.js
@@ -32,16 +32,6 @@ class Book extends Model {
   //   }
   // }
 
-  // static jsonSchema = {
-  //   type: 'object',
-  //   required: ['name'],
-  //   properties: {
-  //     id: { type: 'integer' },
-  //     name: { type: 'string', minLength: 1, maxLength: 255 },
-  //     category: { type: 'string', maxLength: 255 }
-  //   }
-  // }
-
   // This object defines the relations to other models.
   // static relationMappings = {
   static get relationMappings() { 
@@ -49,18 +39,6 @@ class Book extends Model {
     const Page = require('./Page')
     const Category = require('./Category')
     return {
-      // pets: {
-      //   relation: Model.HasManyRelation,
-      //   // The related model. This can be either a Model subclass constructor or an
-      //   // absolute file path to a module that exports one. We use the file path version
-      //   // here to prevent require loops.
-      //   modelClass: Animal,
-      //   join: {
-      //     from: 'persons.id',
-      //     to: 'animals.ownerId'
-      //   }
-      // },
-
       category: { // category, categoryId
         relation: Model.HasOneRelation,
         // relation: Model.BelongsToOneRelation,