+++ example-app/models/User.js
@@ -1,6 +1,5 @@
 'use strict'
 const Model = require('@es-labs/node/services/db/objection').get()
-
 class User extends Model {
   static get tableName () {
     return 'users'