+++ example-webpack/src/layouts/Default.vue
@@ -1,7 +1,0 @@
-<template>
-  <v-app>
-    <v-main>
-      <router-view :key="$route.fullPath"></router-view>
-    </v-main>
-  </v-app>
-</template>