+++ example-webpack/src/permissions.js
@@ -1,6 +1,0 @@
-const permissions = {
-    allowed: (g1, g2) => g1.find(x => g2.includes(x)) // get first group that matches
-}
-  
-export default permissions
-